Added the ability to slide images
Modifications have been made to PropertyCard so that it can now display a slideshow for as long as the length of a given array.
Also, previously the cards displayed on the map and the cards displayed on the side screen used different UIs, but these have now been modified to reference the same UI.
Six images have also been added to display for testing purposes.